Convert startRecording to .NET
