+++ ui-library-add-call-recording/.vscode/extensions.json
@@ -1,3 +1,6 @@
 {
-  "recommendations": ["ms-azuretools.vscode-azurefunctions"]
+  "recommendations": [
+    "ms-azuretools.vscode-azurefunctions",
+    "ms-dotnettools.csharp"
+  ]
 }
