+++ ui-library-add-call-recording/.vscode/launch.json
@@ -2,11 +2,10 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "Attach to Node Functions",
-      "type": "node",
+      "name": "Attach to .NET Functions",
+      "type": "coreclr",
       "request": "attach",
-      "port": 9229,
-      "preLaunchTask": "func: host start"
+      "processId": "${command:azureFunctions.pickProcess}"
     }
   ]
 }
