--- ui-library-add-call-recording/api/.funcignore
@@ -1,7 +1,0 @@
-*.js.map
-*.ts
-.git*
-.vscode
-local.settings.json
-test
-tsconfig.json
