--- ui-library-add-call-recording/api/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "api",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "prestart": "npm run build",
-    "start": "func start",
-    "test": "echo \"No tests yet...\""
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@azure/functions": "^1.2.3",
-    "typescript": "^4.0.0"
-  }
-}
