+++ ui-library-add-call-recording/api/startRecording.cs
@@ -1,0 +1,63 @@
+using System;
+using System.IO;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Extensions.Http;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
+
+namespace Contoso
+{
+    public static class startRecording
+    {
+        [FunctionName("startRecording")]
+        public static async Task<IActionResult> Run(
+            [HttpTrigger(AuthorizationLevel.Anonymous, "get", "post", Route = null)] HttpRequest req,
+            ILogger log)
+        {
+            log.LogInformation("C# HTTP trigger function processed a request.");
+
+            string name = req.Query["name"];
+
+            string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
+            dynamic data = JsonConvert.DeserializeObject(requestBody);
+            name = name ?? data?.name;
+
+            string responseMessage = string.IsNullOrEmpty(name)
+                ? "This HTTP triggered function executed successfully. Pass a name in the query string or in the request body for a personalized response."
+                : $"Hello, {name}. This HTTP triggered function executed successfully.";
+
+            return new OkObjectResult(responseMessage);
+        }
+    }
+}
+
+/*
+const httpTrigger: AzureFunction = async function (
+  context: Context,
+  req: HttpRequest
+): Promise<void> {
+  const callId = req.body?.callId;
+  context.log(`Processing startRecording(${callId})`);
+  if (!callId) {
+    context.res = {
+      // Bad request
+      status: 400,
+      body: "callId is required",
+    };
+    return;
+  }
+
+  context.res = {
+    status: 200,
+    body: JSON.stringify({
+      text: `Would have started recording for ${callId}`,
+    }),
+  };
+};
+
+export default httpTrigger;
+
+*/
