--- ui-library-add-call-recording/api/startRecording/sample.dat
@@ -1,3 +1,0 @@
-{
-    "callId": "dummy-call-id"
-}
