--- ui-library-add-call-recording/api/stopRecording/sample.dat
@@ -1,3 +1,0 @@
-{
-    "callId": "dummy-call-id"
-}
